# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MarvelousBlog-Backend Swagger接口文档",
        "title": "PP同学接口文档",
        "contact": {
            "name": "PP同学",
            "email": "710955321@qq.com"
        },
        "license": {},
        "version": "0.1"
    },
    "paths": {
...
        "/v1/author/login": {
            "post": {
                "description": "输入用户名密码以登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录接口"
                ],
                "summary": "author登录",
                "parameters": [
                    {
                        "description": "作者登录信息",
                        "name": "visitor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginVo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "403": {
                        "description": "用户名或密码错误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "404": {
                        "description": "未找到该用户",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/label/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "输入信息来创建一个标签",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Label接口"
                ],
                "summary": "新增Label标签",
                "parameters": [
                    {
                        "description": "新建的标签信息",
                        "name": "label",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Label"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "新增标签成功",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "403": {
                        "description": "标签重复",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/label/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "输入label_id以删除标签",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Label接口"
                ],
                "summary": "删除某个label",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "删除的label的id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "输入参数有误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "402": {
                        "description": "用户未授权",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/label/list": {
            "get": {
                "description": "获取全部label",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Label接口"
                ],
                "summary": "获取全部label",
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Label"
                            }
                        }
                    },
                    "400": {
                        "description": "输入参数有误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "402": {
                        "description": "用户未授权",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "404": {
                        "description": "未找到资源",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/label/modify/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "输入修改的label信息以更新信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Label接口"
                ],
                "summary": "修改label",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "修改的label id参数",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改的标签信息",
                        "name": "label",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Label"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "402": {
                        "description": "用户未授权",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "403": {
                        "description": "Label重复",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "404": {
                        "description": "未找到资源",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.ResponseObject": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "业务响应码（非HTTP码）"
                },
                "message": {
                    "type": "string",
                    "example": "Some message about the code."
                }
            }
        },
...
        "model.Label": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "如何使用Gin框架搭建一个web服务器"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "labeName": {
                    "type": "string",
                    "example": "Gin框架"
                }
            }
        },
        "model.LoginVo": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string",
                    "example": "PP同学"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}